refactor(repeat): share one exported RRULE normalizer across caldav and ticktick
address pr-swarm findings: caldav's normalizeRRule and ticktick's
normalizeTickTickRepeat reimplemented the same prefix-strip/validate
logic — extracted models.NormalizeRRule (reusing the existing
StrToROption parse path). Fixes two ticktick edge cases along the way:
a mixed-case Rrule: prefix survived the strip and silently dropped the
rule, and a leading DTSTART line ended the scan before the RRULE line.
