+++ pkg/caldav/caldav.go
@@ -205,7 +205,7 @@ PRIORITY:` + strconv.Itoa(mapPriorityToCaldav(t.Priority))
 		if t.Repeats != "" {
 			// Normalize before emitting so we never write a malformed RRULE line
 			// into the calendar output (e.g. from imported or legacy-migrated data).
-			if normalized, ok := normalizeRRule(t.Repeats); ok {
+			if normalized, ok := models.NormalizeRRule(t.Repeats); ok {
 				caldavtodos += `
 RRULE:` + normalized
 			}
