+++ pkg/caldav/parsing.go
@@ -28,27 +28,7 @@ import (
 	"code.vikunja.io/api/pkg/utils"
 
 	ics "github.com/arran4/golang-ical"
-	"github.com/teambition/rrule-go"
 )
-
-// normalizeRRule strips an optional RRULE: prefix, validates the value with the
-// rrule parser, and returns it in canonical form. ok is false when the value is
-// not a parseable RRULE, so callers can skip it instead of storing or emitting
-// malformed calendar data.
-func normalizeRRule(raw string) (normalized string, ok bool) {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		return "", false
-	}
-	if strings.HasPrefix(strings.ToUpper(raw), "RRULE:") {
-		raw = strings.TrimSpace(raw[len("RRULE:"):])
-	}
-	opt, err := rrule.StrToROption(raw)
-	if err != nil {
-		return "", false
-	}
-	return opt.RRuleString(), true
-}
 
 var cssColorsToHex map[string]string
 
@@ -387,7 +367,7 @@ func ParseTaskFromVTODO(content string) (vTask *models.Task, err error) {
 	// whole task on validation.
 	var repeats string
 	if rruleProp, ok := task["RRULE"]; ok {
-		if normalized, valid := normalizeRRule(rruleProp.Value); valid {
+		if normalized, valid := models.NormalizeRRule(rruleProp.Value); valid {
 			repeats = normalized
 		} else {
 			log.Debugf("[CalDAV] Ignoring unparseable RRULE %q on imported task %q", rruleProp.Value, uidValue)
