+++ pkg/caldav/parsing_test.go
@@ -25,7 +25,6 @@ import (
 	"code.vikunja.io/api/pkg/log"
 	"code.vikunja.io/api/pkg/models"
 	ics "github.com/arran4/golang-ical"
-	"github.com/stretchr/testify/assert"
 	"gopkg.in/d4l3k/messagediff.v1"
 )
 
@@ -521,30 +520,6 @@ END:VCALENDAR`,
 			if diff, equal := messagediff.PrettyDiff(got, tt.wantVTask); !equal {
 				t.Errorf("ParseTaskFromVTODO()\n gotVTask = %v\n want %v\n diff = %s", got, tt.wantVTask, diff)
 			}
-		})
-	}
-}
-
-func TestNormalizeRRule(t *testing.T) {
-	cases := []struct {
-		name   string
-		raw    string
-		want   string
-		wantOK bool
-	}{
-		{"plain valid", "FREQ=DAILY;INTERVAL=1", "FREQ=DAILY;INTERVAL=1", true},
-		{"strips RRULE prefix", "RRULE:FREQ=DAILY;INTERVAL=1", "FREQ=DAILY;INTERVAL=1", true},
-		{"strips lowercase prefix", "rrule:FREQ=WEEKLY;BYDAY=MO", "FREQ=WEEKLY;BYDAY=MO", true},
-		{"trims whitespace", "  FREQ=DAILY;INTERVAL=1  ", "FREQ=DAILY;INTERVAL=1", true},
-		{"canonicalizes leading zero", "FREQ=MONTHLY;BYMONTHDAY=01", "FREQ=MONTHLY;BYMONTHDAY=1", true},
-		{"rejects garbage", "not a rule", "", false},
-		{"rejects empty", "", "", false},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			got, ok := normalizeRRule(c.raw)
-			assert.Equal(t, c.wantOK, ok)
-			assert.Equal(t, c.want, got)
 		})
 	}
 }
