+++ pkg/modules/migration/ticktick/ticktick.go
@@ -33,7 +33,6 @@ import (
 	"code.vikunja.io/api/pkg/utils"
 
 	"github.com/gocarina/gocsv"
-	"github.com/teambition/rrule-go"
 )
 
 var timeFormats = []string{
@@ -177,31 +176,20 @@ func sortParentsBeforeChildren(tasks []*tickTickTask) []*tickTickTask {
 }
 
 // normalizeTickTickRepeat normalizes the TickTick Repeat field to a single RRULE string.
-// TickTick exports repeat rules in RRULE format, but may include multiple rules separated by newlines.
-// Vikunja only supports one RRULE, so we take the first one.
+// TickTick exports repeat rules in RRULE format, possibly across multiple lines with a
+// leading DTSTART. Vikunja only supports one RRULE, so we skip DTSTART lines and take
+// the first parseable RRULE line.
 func normalizeTickTickRepeat(repeat string) string {
-	repeat = strings.TrimSpace(repeat)
-	if repeat == "" {
+	for _, line := range strings.Split(repeat, "\n") {
+		line = strings.TrimSpace(line)
+		if line == "" || strings.HasPrefix(strings.ToUpper(line), "DTSTART") {
+			continue
+		}
+		if normalized, ok := models.NormalizeRRule(line); ok {
+			return normalized
+		}
 		return ""
 	}
-
-	// TickTick may have multiple RRULE lines, take the first one
-	lines := strings.Split(repeat, "\n")
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line != "" {
-			// Remove RRULE: prefix if present (TickTick may or may not include it)
-			if strings.HasPrefix(strings.ToUpper(line), "RRULE:") {
-				line = strings.TrimPrefix(line, "RRULE:")
-				line = strings.TrimPrefix(line, "rrule:")
-			}
-			if _, err := rrule.StrToRRule(line); err != nil {
-				return ""
-			}
-			return line
-		}
-	}
-
 	return ""
 }
