+++ pkg/modules/migration/ticktick/ticktick_test.go
@@ -881,6 +881,8 @@ func TestNormalizeTickTickRepeat(t *testing.T) {
 		{"simple RRULE", "FREQ=DAILY;INTERVAL=1", "FREQ=DAILY;INTERVAL=1"},
 		{"RRULE with RRULE prefix", "RRULE:FREQ=WEEKLY;BYDAY=MO", "FREQ=WEEKLY;BYDAY=MO"},
 		{"RRULE with lowercase prefix", "rrule:FREQ=MONTHLY;INTERVAL=2", "FREQ=MONTHLY;INTERVAL=2"},
+		{"RRULE with mixed-case prefix", "Rrule:FREQ=MONTHLY;INTERVAL=2", "FREQ=MONTHLY;INTERVAL=2"},
+		{"DTSTART line before RRULE", "DTSTART:20190101T090000Z\nRRULE:FREQ=DAILY;INTERVAL=1", "FREQ=DAILY;INTERVAL=1"},
 		{"multiline takes first", "FREQ=DAILY;INTERVAL=1\nFREQ=WEEKLY;INTERVAL=2", "FREQ=DAILY;INTERVAL=1"},
 		{"multiline with empty first line", "\nFREQ=DAILY;INTERVAL=1", "FREQ=DAILY;INTERVAL=1"},
 		{"RRULE with UNTIL", "FREQ=WEEKLY;INTERVAL=1;UNTIL=20190117T210000Z", "FREQ=WEEKLY;INTERVAL=1;UNTIL=20190117T210000Z"},
